% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{damr_example}
\alias{damr_example}
\title{Get path to damr example}
\usage{
damr_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
damr comes with a sample DAM2 files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\examples{
damr_example()
damr_example("M014.txt")
}
\author{
Hadley Wickham  (copied from readr)
}
