% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-dam2.R
\name{load_dam2}
\alias{load_dam2}
\title{Load DAM2 data from one or several continuous files}
\usage{
load_dam2(metadata, FUN = NULL, ...)
}
\arguments{
\item{metadata}{\link[data.table:data.table]{data.table::data.table} used to load data (see detail)}

\item{FUN}{function (optional) to transform the data from each animal
immediately after is has been loaded.}

\item{...}{extra arguments to be passed to \code{FUN}}
}
\value{
A \link[behavr:behavr]{behavr::behavr} table.
In addition to the metadata, it contains the data, whith the columns:
\itemize{
\item \code{id} -- autogenerated unique identifier, one per animal
\item \code{t} -- time
\item \code{activity} -- number of beam crosses
}
}
\description{
Uses "linked metadata" to load data from DAM2 arrays.
}
\details{
the linked metadata should be generated using \link{link_dam2_metadata}.
}
\examples{
# This is where our data lives
root_dir <- damr_example_dir()

# Metadata already made for us.
# It defines condition and genotype of each animal
data(single_file_metadata)
print(single_file_metadata)

metadata <- link_dam2_metadata(single_file_metadata, root_dir)

# we find and load the matching data
dt <- load_dam2(metadata)
print(dt)

# genotype and condition to our metadata:
print(dt[meta=TRUE])

# Just the first few reads, we run `head()` on each animal
dt <- load_dam2(metadata, FUN=head)
print(dt)
}
\references{
\itemize{
\item \href{https://rethomics.github.io/behavr.html}{behavr tutorial} -- how to work with the obtained \link{behavr} table
}
}
\seealso{
\itemize{
\item \link[behavr:behavr]{behavr::behavr} --  the class of the resulting object
\item \link{read_dam2_file} --  to load data from a single file (without metadata)
}
}
