% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-dam2.R
\name{query_dam2}
\alias{query_dam2}
\title{Retrieves DAM2 data from one or several continuous files}
\usage{
query_dam2(query, FUN = NULL, ...)
}
\arguments{
\item{query}{\link{data.table} representing a formatted query used to request data (see detail).}

\item{FUN}{function (optional) to transform the data from each animal
immediately after is has been loaded.}

\item{...}{extra arguments to be passed to \code{FUN}}
}
\value{
A \link{behavr} table.
The metadata contains all the query columns and an autogenerated id per animal.
The data has the columns:
\itemize{
\item \code{id} -- autogenerated unique identifier, one per animal
\item \code{t} -- time
\item \code{activity} -- number of beam crosses
}
}
\description{
Uses a query mechanism to get data from a DAM2 array.
This is useful when using the default behaviour of TriKinetics software
where data is simply appended to a single long file per monitor.
}
\details{
\code{query} must be a \link{data.frame} (or \link{data.table}) object.
Conceptually, each row of the query describes one animal with one set of conditions (when \code{region_id} is specified),
or in each monitor (when it is not).
It must have  the following columns:
\itemize{
\item \code{path} -- the location of a data file (e.g. "C:/User/me/Desktop/Monitor3.txt").
\item \code{start_datetme} -- the first day and time of the requested experiment (e.g. "2014-12-28 18:00:00").
\item \code{stop_datetime} -- the last day and time of the requested experiment (e.g. "2014-12-30  19:00:00" or simply "2014-12-30").
\item \code{region_id} -- the channel (between 1 and 32) in which the animal was in (e.g. "20").
\code{region_id} is optional. If not provided, all 32 channels are loaded \emph{with the same conditions}.
\item \code{???} \emph{any number of arbitrary columns} to associate \code{conditions}/\code{treatments}/\code{genotypes}/... to the previous columns.
The time in data is expressed relatively to start_date.
In other words, if you do circadian analysis, and your D-L transitions are at 10:00:00, you want to set
\code{start_datetime = "YYY-MM-DD 10:00:00"}.
}
}
\examples{
sample_file <- damr_example("M064.txt")
query = data.table::data.table(path=sample_file,
                    # note the time (10:00) is added as reference time
                    start_datetime = c("2017-07-01 10:00:00", "2017-07-02 10:00:00"),
                    stop_datetime = "2017-07-07",
                    region_id=c(1:32),
                    condition=rep(letters[1:2],each=16),
                    genotype=c("A", "A", "B", "B"))

dt <- query_dam2(query)
print(dt)
# genotype was added to our metadata:
print(dt[meta=TRUE])
# Just the first few reads, we run head() on each animal
dt <- query_dam2(query, FUN=head)
print(dt)
}
\seealso{
\link{read_dam2_file} to to load data from a single file (without a query).
}
