% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_file_metadata}
\alias{single_file_metadata}
\title{A simple toy metadata defining the experimental conditions of
32 animals monitored at the same time in a single DAM2 monitor.
Each animal has its own channel (region_id), as well as a condition and genotype.
It serves as an example for \link{link_dam_metadata}.}
\format{An object of class \code{data.frame} with 32 rows and 6 columns.}
\usage{
single_file_metadata
}
\description{
A simple toy metadata defining the experimental conditions of
32 animals monitored at the same time in a single DAM2 monitor.
Each animal has its own channel (region_id), as well as a condition and genotype.
It serves as an example for \link{link_dam_metadata}.
}
\author{
Quentin Geissmann
}
\keyword{datasets}
