% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_file_query}
\alias{single_file_query}
\title{A simple toy query defining the experimental conditions of
32 animals monitored at the same time in a single DAM2 monitor.
Each animal has its own channel (region_id), as well as a condition and genotype.
It serves as an example for \link{query_dam2}.}
\format{An object of class \code{data.frame} with 32 rows and 6 columns.}
\usage{
single_file_query
}
\description{
A simple toy query defining the experimental conditions of
32 animals monitored at the same time in a single DAM2 monitor.
Each animal has its own channel (region_id), as well as a condition and genotype.
It serves as an example for \link{query_dam2}.
}
\author{
Quentin Geissmann
}
\keyword{datasets}
