% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_files_metadata}
\alias{two_files_metadata}
\title{A toy metadata defining the experimental conditions of
64 animals monitored at the same time in two separate DAM2 monitors.
Each animal has its own channel (region_id), as well as a condition and genotype.
It serves as an example for \link{link_dam_metadata}.}
\format{
An object of class \code{data.frame} with 64 rows and 6 columns.
}
\usage{
two_files_metadata
}
\description{
A toy metadata defining the experimental conditions of
64 animals monitored at the same time in two separate DAM2 monitors.
Each animal has its own channel (region_id), as well as a condition and genotype.
It serves as an example for \link{link_dam_metadata}.
}
\author{
Quentin Geissmann
}
\keyword{datasets}
