% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{damr_example}
\alias{damr_example}
\alias{damr_example_dir}
\title{Get path to damr example}
\usage{
damr_example(path = NULL)

damr_example_dir()
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
damr comes with a sample DAM2 files in its \code{inst/extdata}
directory. \code{damr_example} allow  make them easy to access.
}
\examples{
# list all files
damr_example()
# get path to one file
damr_example("M014.txt")
# get the directory wih all the files
damr_example_dir()
}
\author{
Hadley Wickham  (modified from readr)
}
